requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Flask-JWT-Extended
pytest-flask
flask-bcrypt
marshmallow
flask-swagger-ui
flask-restx